+++ package.json
@@ -40,6 +40,7 @@
   },
   "devDependencies": {
     "@types/react": "^18.2.43",
+    "@types/react-color": "^3.0.11",
     "@types/react-dom": "^18.2.17",
     "@typescript-eslint/eslint-plugin": "^6.14.0",
     "@typescript-eslint/parser": "^6.14.0",
