+++ src/Router.tsx
@@ -3,13 +3,13 @@ import { createBrowserRouter } from 'react-router-dom';
 import Start from './pages/Start';
 import Login from './pages/Login';
 import Signup from './pages/Signup';
-import Createstyle from './pages/Createstyle';
 import MainLayout from './layout/MainLayout';
 import All from './pages/AllForm';
 import MyForm from './pages/MyForm';
 import MyResponse from './pages/MyResponse';
-import ResponseForm from './pages/ResponseForm';
+import Create from './pages/Create';
 import ResultPage from './pages/ResultPage';
+import ResponseForm from './pages/ResponseForm';
 
 interface RouterElement {
   id: number;
@@ -26,7 +26,7 @@ const routerData: RouterElement[] = [
   { id: 4, path: '/all', element: <All />, withAuth: false, isNavbar: true },
   { id: 5, path: '/myform', element: <MyForm />, withAuth: false, isNavbar: true },
   { id: 6, path: '/myresponses', element: <MyResponse />, withAuth: false, isNavbar: true },
-  { id: 7, path: '/createstyle', element: <Createstyle />, withAuth: false, isNavbar: true },
+  { id: 7, path: '/create', element: <Create />, withAuth: false, isNavbar: true },
   { id: 8, path: '/result/:id', element: <ResultPage />, withAuth: true, isNavbar: true },
   { id: 9, path: '/responseform', element: <ResponseForm />, withAuth: false, isNavbar: true },
 ];
